Various Fixes, New Binary Format etc.
New binary format added in. Used in Undo / Redo system. All future graphs will be saved in binary format. Old graph format is still supported but expect slower loading. Binary format not as tested yet as the previous JSON format. Need to test with all nodes etc and make sure it produces the proper results.

Todo: rework one enum conversion in binary write / read to another where it is using TyrAndParse, and instead just do a switch and manual assignment.

Fixed issue with undo redo system, also added in a differential restore for updating the nodes in the graph after an undo / redo. Previously it was disposing everything and recreating everything. That was totally unnecessary.

Internal value changes to some stuff, such as going from int to ushort etc. To better match actual allowed values for binary format.

Removed stuff related to old undo / redo system that was no longer needed.

Fixed a bug in Graph\IO\Reader where it was calling NextInt() for NextBool() when it should have been NextByte().

Added in UIResizable component to handle resizing of elements with pointer / mouse.

removed unneeded calls to various stencilmask in GL and etc in UI.

Fixed an issue where UIImage AdjustStencil was only being called if it was not Clipped. It is now called in all cases properly.

Working on main menu implementation and documents handler / tabs. Not fully complete yet.

PBR shader was incorrectly flipping all textures UV. This was left over from previous implementation, and is no longer needed. Since the textures by default are all in the proper  GL orientation.

Made Hdri manager more self contained for loading and processing.

Todo: on main menu update save / save as / save all / close all  enabled states, add in proper recent data source, etc.

+++ Graph/Image.cs
@@ -9,12 +9,12 @@ namespace Materia.Graph
 {
     public class Image : Graph
     {
-        public Image(string name, int w = 256, int h = 256) : base (name, w, h)
+        public Image(string name, ushort w = 256, ushort h = 256) : base (name, w, h)
         {
 
         }
 
-        public Image(string name, int w = 256, int h = 256, Graph parent = null) : base(name, w, h)
+        public Image(string name, ushort w = 256, ushort h = 256, Graph parent = null) : base(name, w, h)
         {
             parentGraph = parent;
         }
